task10

+++ 0x05-pointers_arrays_strings/100-print_number.c
@@ -7,9 +7,9 @@
  */
 void print_number(int n)
 {
-	unsigned long expo;
-	unsigned long temp;
-	unsigned long temp2;
+	unsigned int expo;
+	unsigned int temp;
+	unsigned int temp2;
 
 	expo = 1;
 	temp2 = n;
